+++ ReactApp/src/App.js
@@ -11,6 +11,7 @@ import AppBar from 'material-ui/AppBar';
 import IndexPage from './Pages/Index';
 import ProgramPage from './Pages/Program';
 import ActivitiesPage from './Pages/Activities';
+import LuthiersPage from './Pages/Luthiers';
 import ArtisansPage from './Pages/Artisans';
 import FoodPage from './Pages/Food';
 import TicketsPage from './Pages/Tickets';
@@ -23,6 +24,7 @@ const pagesMap = {
   index: IndexPage,
   program: ProgramPage,
   artisans: ArtisansPage,
+  luthiers: LuthiersPage,
   food: FoodPage,
   activities: ActivitiesPage,
   tickets: TicketsPage,
